Day9

+++ src/main/java/com/callibrity/adventofcode/intcode/ops/Multiplication.java
@@ -3,8 +3,8 @@ package com.callibrity.adventofcode.intcode.ops;
 public class Multiplication implements Operation {
     @Override
     public void execute(OperationContext context) {
-        final int left = context.nextParameter();
-        final int right = context.nextParameter();
+        final long left = context.nextParameter();
+        final long right = context.nextParameter();
         context.storeValue(left * right);
     }
 }
